Refactoring in the code.

+++ spark/src/main/java/com/talentica/hungryHippos/rdd/main/SumJobTextFile.java
@@ -61,7 +61,7 @@ public class SumJobTextFile implements Serializable {
         cacheRDD.put(keyOfHHRDD, hipposRDD);
       }
       Broadcast<Job> jobBroadcast = context.broadcast(job);
-      JavaRDD<Tuple2<String, Double>> resultRDD =
+      JavaRDD<Tuple2<String, Integer>> resultRDD =
           new SumExecutor<String>().process(hipposRDD, descriptionBroadcast, jobBroadcast);
       String outputDistributedPath =
           outputDirectory + File.separator + jobBroadcast.value().getJobId();
